# Conflicts:
#	lucene/common-build.xml
#	lucene/ivy-versions.properties
#	solr/bin/solr
#	solr/contrib/uima/build.xml
#	solr/core/src/java/org/apache/solr/core/CoreContainer.java
#	solr/core/src/java/org/apache/solr/util/SolrCLI.java
#	solr/server/solr/configsets/_default/conf/solrconfig.xml
#	solr/test-framework/src/java/org/apache/solr/SolrTestCaseJ4.java
#	solr/webapp/web/index.html
#	solr/webapp/web/js/angular/app.js
#	solr/webapp/web/js/angular/services.js

+++ lucene/core/src/java/org/apache/lucene/util/fst/BytesStore.java
@@ -77,10 +77,10 @@ class BytesStore extends DataOutput implements Accountable {
 
   /** Absolute write byte; you must ensure dest is &lt; max
    *  position written so far. */
-  public void writeByte(int dest, byte b) {
-    int blockIndex = dest >> blockBits;
+  public void writeByte(long dest, byte b) {
+    int blockIndex = (int) (dest >> blockBits);
     byte[] block = blocks.get(blockIndex);
-    block[dest & blockMask] = b;
+    block[(int) (dest & blockMask)] = b;
   }
 
   @Override
